requirements.txt
asgiref==3.3.4
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.2.3
django-annoying==0.10.6
django-cors-headers==3.7.0
django-thumbs-v2==0.4.1
djangorestframework==3.12.4
gunicorn==20.1.0
mysqlclient==2.0.3
numpy==1.20.3
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
whitenoise==5.2.0
